requirements.txt
transformers
torch
nltk
flask
requests
beautifulsoup4
newspaper3k
lxml
lxml_html_clean
json
flask